--- packages/client/src/game/GunState.ts
@@ -1,6 +1,0 @@
-enum GunState {
-  None,
-  Carrying,
-  Held,
-}
-export default GunState;
