--- packages/client/src/game/MouseState.ts
@@ -1,8 +1,0 @@
-enum MouseState {
-  None,
-  Place,
-  Erase,
-  WasPlacingNowErasing,
-  WasErasingNowPlacing,
-}
-export default MouseState;
