--- packages/client/src/game/Network.ts
@@ -1,5 +1,0 @@
-import Game from "./Game";
-
-export default interface Network {
-  update(game: Game): void;
-}
