+++ packages/client/src/game/client/ClientGame.ts
@@ -2,13 +2,13 @@ import type { ZSInit } from "@smiley-face-game/api/packets";
 import type { Connection } from "@smiley-face-game/api";
 import type { Renderer } from "pixi.js";
 import type TileRegistration from "@smiley-face-game/api/tiles/TileRegistration";
-import Game from "./Game";
-import Bullets from "./Bullets";
-import Chat from "./Chat";
+import Game from "../Game";
+import Bullets from "../Bullets";
+import Chat from "../Chat";
 import ClientPlayers from "./ClientPlayers";
 import ClientWorld from "./ClientWorld";
 import ClientNetwork from "./ClientNetwork";
-import ClientBullet from "./ClientBullet";
+import ClientBullet from "./components/ClientBullet";
 import AuthoredBlockPlacer from "./AuthoredBlockPlacer";
 import ClientSelector from "./ClientSelector";
 import ClientDisplay from "./ClientDisplay";
