+++ packages/client/src/game/client/Keyboard.ts
@@ -1,4 +1,4 @@
-import Player from "./Player";
+import Player from "../components/Player";
 
 export default class Keyboard {
   constructor(player: Player) {
