+++ packages/client/src/game/helpers/defaultInputs.ts
@@ -1,4 +1,4 @@
-import Inputs from "./Inputs";
+import Inputs from "../interfaces/Inputs";
 
 export default function defaultInputs(): Inputs {
   return { jump: false, up: false, down: false, left: false, right: false };
