+++ packages/client/src/game/helpers/newCompositeRectTileLayer.ts
