+++ packages/client/src/game/interfaces/BulletCtor.ts
@@ -1,4 +1,4 @@
-import Bullet from "./Bullet";
+import Bullet from "../components/Bullet";
 
 export default interface BulletCtor {
   new (x: number, y: number, angle: number): Bullet;
