+++ packages/client/src/game/interfaces/Display.ts
@@ -1,4 +1,4 @@
-import Game from "./Game";
+import Game from "../Game";
 
 export default interface Display {
   draw(game: Game): void;
