+++ packages/client/src/game/interfaces/PhysicsObject.ts
@@ -1,6 +1,6 @@
 import Position from "./Position";
 import Velocity from "./Velocity";
-import Game from "./Game";
+import Game from "../Game";
 
 export default interface PhysicsObject {
   position: Position;
