+++ packages/client/src/game/interfaces/PlayerCtor.ts
@@ -1,4 +1,4 @@
-import Player from "./Player";
+import Player from "../components/Player";
 
 export default interface PlayerCtor {
   new (id: number, username: string, isGuest: boolean): Player;
