+++ packages/client/src/game/interfaces/ReadOnlyWorld.ts
