+++ packages/client/src/ui/pages/PlayPage.jsx
@@ -20,7 +20,7 @@ import MobileControls from "../game/MobileControls";
 import WorldSettings from "../game/WorldSettings";
 import { Authentication } from "@smiley-face-game/api";
 import { Renderer } from "pixi.js";
-import makeClientConnectedGame from "../../game/makeClientConnectedGame";
+import makeClientConnectedGame from "../../game/helpers/makeClientConnectedGame";
 import textures from "../../game/textures";
 
 const useStyles = makeStyles({
